psydaily/requirements.txt
feedparser>=6.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-telegram-bot>=20.0
schedule>=1.2.0
python-dotenv>=1.0.0
